Extract common dialog code
Since many dialogs will have an OK & Cancel button, and
may not want to inherit from `simpledialog` in future, extract
common code into `OkCancelDialog` - currently derived from
`simpledialog` but not necessarily in future.

Other common dialog types can be added to `dialogs.py` as
needed.
