+++ docs/index.rst
@@ -36,3 +36,8 @@ Page Details
 ============
     
 .. automodule:: page_details
+ 
+Dialogs
+=======
+    
+.. automodule:: dialogs
